pyproject.toml
[project]
name = "mom-missing-charters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.27.1",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
